requirements.txt
torch
torchvision
torchaudio
facenet-pytorch
opencv-python
numpy
scikit-learn
Pillow
matplotlib
onnx
onnxruntime
requests